hopfully successfully implementing logs every n steps

+++ mutformer_model_code/run_pretraining.py
@@ -146,7 +146,7 @@ def model_fn_builder(bert_config, init_checkpoint, init_learning_rate,
                                 return tf.contrib.summary.all_summary_ops()
                     def not_writing():
                         return []
-                    return tf.cond(tf.equal(tf.mod(gs,tf.constant(save_logs_every_n_steps)),tf.constant(0)),writing,not_writing)
+                    return tf.cond(tf.equal(tf.mod(tf.cast(gs,tf.int32),tf.constant(save_logs_every_n_steps)),tf.constant(0)),writing,not_writing)
                 gs_t = tf.reshape(global_step, [1])
                 loss_t = tf.reshape(total_loss, [1])
                 lr_t = tf.reshape(learning_rate, [1])
